+++ plugin/src/main/scala/com/stratio/cassandra/lucene/partitioning/Partitioner.scala
@@ -79,7 +79,8 @@ object Partitioner {
   @JsonSubTypes(Array(
     new JsonSubTypes.Type(value = classOf[PartitionerOnNone.Builder], name = "none"),
     new JsonSubTypes.Type(value = classOf[PartitionerOnToken.Builder], name = "token"),
-    new JsonSubTypes.Type(value = classOf[PartitionerOnColumn.Builder], name = "column")))
+    new JsonSubTypes.Type(value = classOf[PartitionerOnColumn.Builder], name = "column"),
+    new JsonSubTypes.Type(value = classOf[PartitionerOnVNodes.Builder], name = "vnodes")))
   trait Builder {
 
     def build(metadata: CFMetaData): Partitioner